HW7/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
requests==2.22.0
requests-oauthlib==1.3.0